Write a program to count the number of words in a text file  and upload it to GitHub.
task - 14 Create a file reader utility tool and upload it to GitHub. Share how this helped you understand Java's IO Streams. Create a Student class with fields name and marks. Serialize and deserialize an object of this class to and from a file. Write a program to count the number of words in a text file.