Create an interface Shape with an abstract method calculateArea(). Create an abstract class Rectangle that implements Shape. It should have two protected instance variables: width and height. It should have a constructor that takes width and height as arguments. Create an anonymous class that extends Rectangle and provides the implementation for the calculateArea() method. Create a nested static class Circle within the Rectangle class that implements Shape. It should have a private instance variable radius. It should have a constructor that takes radius as an argument. It should provide the implementation for the calculateArea() method. Create a main class to: Create an instance of the anonymous Rectangle class. Create an instance of the Rectangle.Circle class. Print the areas of both shapes.
Task: 19 Create an interface Shape with an abstract method calculateArea(). Create an abstract class Rectangle that implements Shape. It should have two protected instance variables: width and height. It should have a constructor that takes width and height as arguments. Create an anonymous class that extends Rectangle and provides the implementation for the calculateArea() method. Create a nested static class Circle within the Rectangle class that implements Shape. It should have a private instance variable radius. It should have a constructor that takes radius as an argument. It should provide the implementation for the calculateArea() method. Create a main class to: Create an instance of the anonymous Rectangle class. Create an instance of the Rectangle.Circle class. Print the areas of both shapes.