Given a list of integers, use Predicate<Integer> and a lambda expression to filter out all numbers greater than 10. Given a list of strings, use Function<String, String> and a lambda expression to convert each string to uppercase. Given a list of strings, use Consumer<String> and a lambda expression to print each string in uppercase to the console. sort a list of Person objects based on the second character of their names. Using lambda expression using comparator.
task 32: 
Given a list of integers, use Predicate<Integer> and a lambda expression to filter out all numbers greater than 10.
Given a list of strings, use Function<String, String> and a lambda expression to convert each string to uppercase.
Given a list of strings, use Consumer<String> and a lambda expression to print each string in uppercase to the console.
sort a list of Person objects based on the second character of their names. Using lambda expression using comparator.